+++ javascope/NotConnectedDataProvider.java
@@ -37,7 +37,7 @@ class NotConnectedDataProvider implements DataProvider
     public void SetEnvironment(String exp)
     {
     }
-    public void Update(String exp, int s)
+    public void Update(String exp, long s)
     {
     }
     public String GetString(String in)
@@ -60,9 +60,9 @@ class NotConnectedDataProvider implements DataProvider
         return null;
     }
     
-    public int[] GetShots(String in)
+    public long[] GetShots(String in)
     {
-        int d[] = new int[1];
+        long d[] = new long[1];
         d[0] = 0;
         return d;        
     }
